M2.5 follow-ups: batch dev in expansion, VAE fix, dev knobs everywhere, per-image timing
- Batch dev: run_jobs routes flux.2-dev to the quantized Comfy loaders (TE/transformer/VAE),
  un-guarded for the Stage-B coverage sweep; adapter carries text_encoder/fp8_matmul in the
  batch shared block. Stage-B endpoint was already dev-aware (JSON prompts, 512).
- VAE dtype fix: load the dev VAE in float32 (Klein parity) so ref/i2i encode of a float32
  image works (the spike's bf16 VAE only ever ran t2i). Regression test added.
- FE: dev knobs (text_encoder/fp8_matmul) render inline on every flux2 params surface; fixed
  Stage-B visibility (ParamControls gains a modelName prop, since Stage-B holds its model in
  stageBModel, not values.model_name).
- Inspector: per-image generation time - _batch + multi adapters propagate per-cell/
  per-candidate duration_s into output_meta; inspector shows batch + image for batch jobs.
- Tests + journal updated. Suite 326 green; tsc + vite build clean.

+++ pipelines/multistack/src/pipeline/flux2/stage1_load_models.py
@@ -137,7 +137,11 @@ def _load_dev_quantized(
     timings["flow_model_load_s"] = round(time.time() - t0, 4)
 
     t0 = time.time()
-    ae, vae_stats = q.load_comfy_vae(vae_path, device=torch_device, dtype=torch_dtype)
+    # The VAE runs in float32 (matching Klein's load_ae) — NOT bf16 like the transformer/TE. The
+    # spike used bf16 here, which works for t2i (decode only) but breaks ref/i2i: encode_image_refs
+    # feeds a float32 image into conv_in, and decode's inv_normalize (float32 bn buffers) promotes
+    # the bf16 latent to float32 anyway. So float32 is the dtype the encode + decode both need.
+    ae, vae_stats = q.load_comfy_vae(vae_path, device=torch_device, dtype=torch.float32)
     timings["autoencoder_load_s"] = round(time.time() - t0, 4)
 
     return {
